Move ForkChoiceUtil into core package
