+++ protoarray/build.gradle
@@ -1,7 +1,6 @@
 dependencies {
     implementation 'org.apache.tuweni:tuweni-crypto'
 
-    implementation project(':storage')
     implementation project(':ethereum:datastructures')
     implementation project(':ssz')
     implementation project(':util')
