+++ protoarray/src/main/java/tech/pegasys/artemis/protoarray/ForkChoiceStrategy.java
@@ -16,12 +16,12 @@ package tech.pegasys.artemis.protoarray;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.artemis.datastructures.blocks.BeaconBlock;
 import tech.pegasys.artemis.datastructures.forkchoice.MutableStore;
+import tech.pegasys.artemis.datastructures.forkchoice.ReadOnlyStore;
 import tech.pegasys.artemis.datastructures.operations.IndexedAttestation;
-import tech.pegasys.artemis.storage.Store;
 
 public interface ForkChoiceStrategy {
 
-  Bytes32 findHead(final Store store);
+  Bytes32 findHead(final ReadOnlyStore store);
 
   void onAttestation(final IndexedAttestation attestation);
