+++ protoarray/src/testFixtures/java/tech/pegasys/artemis/protoarray/StubForkChoiceStrategy.java
@@ -16,13 +16,13 @@ package tech.pegasys.artemis.protoarray;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.artemis.datastructures.blocks.BeaconBlock;
 import tech.pegasys.artemis.datastructures.forkchoice.MutableStore;
+import tech.pegasys.artemis.datastructures.forkchoice.ReadOnlyStore;
 import tech.pegasys.artemis.datastructures.operations.IndexedAttestation;
-import tech.pegasys.artemis.storage.Store;
 
 public class StubForkChoiceStrategy implements ForkChoiceStrategy {
 
   @Override
-  public Bytes32 findHead(final Store store) {
+  public Bytes32 findHead(final ReadOnlyStore store) {
     return Bytes32.ZERO;
   }
